# zstd: Big speedup on small dictionary encodes