--- a/zstd/enc_fast.go
+++ b/zstd/enc_fast.go
@@ -975,6 +975,7 @@ func (e *fastEncoderDict) Reset(d *dict, singleBlock bool) {
 			}
 		}
 		e.lastDictID = d.id
+		e.allDirty = true
 	}
 
 	e.cur = e.maxMatchOff