@tony2001 2021-03-23
According to my benchmarks, the best results are achieved with tableShardSize = 64, (that would mean tableShardCnt  = 1 << (tableBits - 6) in enc_fast.go).
This number is empirical, but I believe the reason is that sizeof(tableEntry) == 64, which means that each "shard" is 64*64=4096 bytes, standard page size on Linux.

Other than that, it seems your version is just my patch on steroids :), so I still get the same results in the benchmarks (after adjusting the shard size, that is).

RE: code beauty
Do you think a table struct with getters/setters doing the "dirty" stuff behind the scenes would affect performance negatively? It just asks for some kind of general solution for all these tables, but my attempts resulted in like -10%-20% of speed =(