@klauspost 2021-03-24
@tony2001 

> tableShardSize = 64

I will make it a tweakable constant and benchmark the numbers. It will mainly be a tradeoff between the cost of branching vs the cost of copying memory. So a system with less memory bandwidth will prefer smaller shards.

> Do you think a table struct with getters/setters doing the "dirty" stuff behind the scenes would affect performance negatively? It just asks for some kind of general solution for all these tables, but my attempts resulted in like -10%-20% of speed =(

I looked into that. The main issue is that the `e.tableShardDirty[entryNum/tableShardSize]` division must be resolvable at compile time, otherwise performance will be horrible. 

Less important, but also important is the `i*shardSize` multiplication which also benefits greatly from being constant. Though, if the shard size is constant for all tables that can be achieved with a generic implementation.