@klauspost 2021-03-24
I would have liked to be able to the same for all like the "fast", where it falls back to the regular encode when exceeding a certain size:

```
	if e.allDirty || len(src) > 32<<10 {
		e.fastEncoder.Encode(blk, src)
		e.allDirty = true
		return
	}
```

.. but I couldn't find a neat way for that.